refactor(orm,module): reorganize bootstrap, install pipeline, and data sync
Rename ui view lookup; merge company ABAC links; split setup validate/seed;
extract install phase functions and action sync; fold view inherit into
data_sync_views; remove view_inherit_load.go.
